Move Storing Purchase Order Logic
