+++ resources/views/vendors/bills/index.blade.php
@@ -355,7 +355,7 @@
                 <p class="h3 pl-4 m-auto">Purchase Order</p>
                 <a class="close" data-dismiss="modal">×</a>
             </div>
-            <form class="ajax-submit-updated" id="form-purchase-order" action="{{route('bills.purchaseOrder.store') }}"
+            <form class="ajax-submit-updated" id="form-purchase-order" action="{{ url('/vendors/bills/purchaseorder') }}"
                 method="post" data-message="Successfully created a purchase order.">
                 @csrf
                 @include('vendors.bills.forms.purchaseOrderModal')
